requirements.txt
python-kraken-sdk>=3.2.0,<4.0.0
pandas>=2.1.0
pandas-ta>=0.3.14b
python-dotenv>=1.0.0
PyYAML>=6.0.1
aiosqlite>=0.20.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4
python-multipart>=0.0.9
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
